Makefile and python script for building added
The makefile for Arduino, based on the makefile by Sudar has been added and modified so it works with our BLE project.

Also a simple python script was written for building all the examples on the BLE folder.
A small Makefile was added on all examplesfor building them. For more details about how ot use it see the attached info file.